[2741] Reorganize structure, remove static imports, improve animations, fix missing background necessary for proper dark mode operation
