+++ stream-chat-android-compose/api/stream-chat-android-compose.api
@@ -153,6 +153,16 @@ public final class io/getstream/chat/android/compose/state/imagepreview/ImagePre
 	public static fun values ()[Lio/getstream/chat/android/compose/state/imagepreview/ImagePreviewResultType;
 }
 
+public final class io/getstream/chat/android/compose/state/messageoptions/MessageOptionItemState {
+	public static final field $stable I
+	public synthetic fun <init> (IJLandroidx/compose/ui/graphics/painter/Painter;JLio/getstream/chat/android/common/state/MessageAction;Lkotlin/jvm/internal/DefaultConstructorMarker;)V
+	public final fun getAction ()Lio/getstream/chat/android/common/state/MessageAction;
+	public final fun getIconColor-0d7_KjU ()J
+	public final fun getIconPainter ()Landroidx/compose/ui/graphics/painter/Painter;
+	public final fun getTitle ()I
+	public final fun getTitleColor-0d7_KjU ()J
+}
+
 public final class io/getstream/chat/android/compose/state/messages/MessageAlignment : java/lang/Enum {
 	public static final field End Lio/getstream/chat/android/compose/state/messages/MessageAlignment;
 	public static final field Start Lio/getstream/chat/android/compose/state/messages/MessageAlignment;
@@ -376,16 +386,6 @@ public abstract class io/getstream/chat/android/compose/state/messages/list/Mess
 	public static final field $stable I
 }
 
-public final class io/getstream/chat/android/compose/state/messages/list/MessageOptionItemState {
-	public static final field $stable I
-	public synthetic fun <init> (IJLandroidx/compose/ui/graphics/painter/Painter;JLio/getstream/chat/android/common/state/MessageAction;Lkotlin/jvm/internal/DefaultConstructorMarker;)V
-	public final fun getAction ()Lio/getstream/chat/android/common/state/MessageAction;
-	public final fun getIconColor-0d7_KjU ()J
-	public final fun getIconPainter ()Landroidx/compose/ui/graphics/painter/Painter;
-	public final fun getTitle ()I
-	public final fun getTitleColor-0d7_KjU ()J
-}
-
 public final class io/getstream/chat/android/compose/state/messages/list/SendGiphy : io/getstream/chat/android/compose/state/messages/list/GiphyAction {
 	public static final field $stable I
 	public fun <init> (Lio/getstream/chat/android/client/models/Message;)V
@@ -420,14 +420,14 @@ public final class io/getstream/chat/android/compose/state/messages/list/ThreadS
 	public fun toString ()Ljava/lang/String;
 }
 
-public final class io/getstream/chat/android/compose/state/messages/reaction/ReactionOptionItemState {
+public final class io/getstream/chat/android/compose/state/reactionoptions/ReactionOptionItemState {
 	public static final field $stable I
 	public fun <init> (Landroidx/compose/ui/graphics/painter/Painter;ZLjava/lang/String;)V
 	public final fun component1 ()Landroidx/compose/ui/graphics/painter/Painter;
 	public final fun component2 ()Z
 	public final fun component3 ()Ljava/lang/String;
-	public final fun copy (Landroidx/compose/ui/graphics/painter/Painter;ZLjava/lang/String;)Lio/getstream/chat/android/compose/state/messages/reaction/ReactionOptionItemState;
-	public static synthetic fun copy$default (Lio/getstream/chat/android/compose/state/messages/reaction/ReactionOptionItemState;Landroidx/compose/ui/graphics/painter/Painter;ZLjava/lang/String;ILjava/lang/Object;)Lio/getstream/chat/android/compose/state/messages/reaction/ReactionOptionItemState;
+	public final fun copy (Landroidx/compose/ui/graphics/painter/Painter;ZLjava/lang/String;)Lio/getstream/chat/android/compose/state/reactionoptions/ReactionOptionItemState;
+	public static synthetic fun copy$default (Lio/getstream/chat/android/compose/state/reactionoptions/ReactionOptionItemState;Landroidx/compose/ui/graphics/painter/Painter;ZLjava/lang/String;ILjava/lang/Object;)Lio/getstream/chat/android/compose/state/reactionoptions/ReactionOptionItemState;
 	public fun equals (Ljava/lang/Object;)Z
 	public final fun getPainter ()Landroidx/compose/ui/graphics/painter/Painter;
 	public final fun getType ()Ljava/lang/String;
@@ -746,7 +746,7 @@ public final class io/getstream/chat/android/compose/ui/components/messageoption
 }
 
 public final class io/getstream/chat/android/compose/ui/components/messageoptions/MessageOptionItemKt {
-	public static final fun MessageOptionItem (Lio/getstream/chat/android/compose/state/messages/list/MessageOptionItemState;Landroidx/compose/ui/Modifier;Landroidx/compose/ui/Alignment$Vertical;Landroidx/compose/foundation/layout/Arrangement$Horizontal;Landroidx/compose/runtime/Composer;II)V
+	public static final fun MessageOptionItem (Lio/getstream/chat/android/compose/state/messageoptions/MessageOptionItemState;Landroidx/compose/ui/Modifier;Landroidx/compose/ui/Alignment$Vertical;Landroidx/compose/foundation/layout/Arrangement$Horizontal;Landroidx/compose/runtime/Composer;II)V
 }
 
 public final class io/getstream/chat/android/compose/ui/components/messageoptions/MessageOptionsKt {
@@ -806,7 +806,7 @@ public final class io/getstream/chat/android/compose/ui/components/messages/Mess
 }
 
 public final class io/getstream/chat/android/compose/ui/components/messages/MessageReactionItemKt {
-	public static final fun MessageReactionItem (Lio/getstream/chat/android/compose/state/messages/reaction/ReactionOptionItemState;Landroidx/compose/ui/Modifier;Landroidx/compose/runtime/Composer;II)V
+	public static final fun MessageReactionItem (Lio/getstream/chat/android/compose/state/reactionoptions/ReactionOptionItemState;Landroidx/compose/ui/Modifier;Landroidx/compose/runtime/Composer;II)V
 	public static final fun MessageReactionItemNotSelectedPreview (Landroidx/compose/runtime/Composer;I)V
 	public static final fun MessageReactionItemSelectedPreview (Landroidx/compose/runtime/Composer;I)V
 }
@@ -860,7 +860,7 @@ public final class io/getstream/chat/android/compose/ui/components/reactionoptio
 }
 
 public final class io/getstream/chat/android/compose/ui/components/reactionoptions/ReactionOptionItemKt {
-	public static final fun ReactionOptionItem (Lio/getstream/chat/android/compose/state/messages/reaction/ReactionOptionItemState;Landroidx/compose/ui/Modifier;Landroidx/compose/runtime/Composer;II)V
+	public static final fun ReactionOptionItem (Lio/getstream/chat/android/compose/state/reactionoptions/ReactionOptionItemState;Landroidx/compose/ui/Modifier;Landroidx/compose/runtime/Composer;II)V
 }
 
 public final class io/getstream/chat/android/compose/ui/components/reactionoptions/ReactionOptionsKt {
