+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/previewdata/PreviewReactionData.kt
@@ -3,7 +3,7 @@ package io.getstream.chat.android.compose.previewdata
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.painterResource
 import io.getstream.chat.android.compose.R
-import io.getstream.chat.android.compose.state.messages.reaction.ReactionOptionItemState
+import io.getstream.chat.android.compose.state.reactionoptions.ReactionOptionItemState
 
 /**
  * Provides sample reactions that will be used to render component previews.
